Implement blockSetExtRequest
While storing new credentials there is a bug that sometimes the
credentials where not properly stored. This is due to a new `ExtRequest`
overriding the one that should still be processed in the background
(e.g., we first set the context, then change the `extRequest` to set the
login and after that we change it again to set the password).

To make sure this sequence of messages to the device is not interrupted
resulting in partially saved credentials, we implement
`blockSetExtRequest`. This should block any new requests being made
while we are still in a sequence.